Bug 13183 - Improve visibility of automatic on-site checkout rule when OnSiteCheckoutsForce is on
This patch addresses the case where OnSiteCheckoutsForce is enabled and
the user is checking out to a patron who is blocked from checking out.
Instead of showing the 'specify due date' form (which is disabled) this
patch hides it and highlights the on-site checkout controls.

This patch adds some CSS and consolidates some redundant CSS.

To test, load a patron for checkout under the following conditions:

1. OnSiteCheckoutsForce is off and patron is not blocked from checkout
2. OnSiteCheckoutsForce is off, patron is blocked from checkout (for
   being expired or restricted).
3. OnSiteCheckoutsForce is on and patron is not blocked from checkout
4. OnSiteCheckoutsForce is on and patron is blocked from checkout

Signed-off-by: Liz Rea <liz@catalyst.net.nz>
Checked these 4 conditions, plus that if a specified due date was allowed, that it was enforced.

I noted that the on-site checkouts seemed to take longer to process, but this probably isn't the place for that sort of complaint.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Small visual improvment for a new feature. No problems found.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>