Bug 13194: Fix typo in SetMessagingPreferencesFromDefaults POD
The example usage for
C4::Members::Messaging::SetMessagingPreferencesFromDefaults
calls the subroutine SetMessagingPreferenceFromDefaults, not
SetMessagingPreferencesFromDefaults (missing the 's' at the
end of 'Preferences').

To test:
- Apply the patch
- Check that the POD now refers to the actual name of the
  subroutine
  (perldoc C4::Members::Messaging)

http://bugs.koha-community.org/show_bug.cgi?id=13194

Adding 's' is the correct doc change to make it match with
the function definition. Comfirmable with less.
perldoc C4::Members::Messaging proves the doc is still nice.

Signed-off-by: Mark Tompsett <mtompset@hotmail.com>

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>