Bug 12059: Publisher column on invoice page always empty
This patch moves the publisher information out of its own
always empty column into the Summary column below the title,
as it is on other acq pages.

The information was never displaying, as publishercode is in
biblioitems and that table was not selected by GetInvoiceDetails.

Also modified the code to take into account that UNIMARC uses
biblioitems.publicationyear and MARC21/NORMARC use bibio.copyrightdate
for the copyright year.

To test:
- create an invoice for records that
  - have a publication year
  - have no publication year
  - have a publisher...
- 'finish receiving' and check the invoice summary page
   ...acqui/invoice.pl?invoiceid=?
- Make sure all the information displays now but didn't witout the patch.

Signed-off-by: Paola Rossi <paola.rossi@cineca.it>

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>