Bug 13271 - Subject search links broken if XSLT and OPACPopupAuthorsSearch are both off
This patch corrects some faults in the OPAC's bootstrap template which
prevent additional author and subject searches from working correctly
with OPACPopupAuthorsSearch both on and off.

Some JavaScript has been removed because it was used only by the prog
theme and is obsolete.

Locate a title with multiple subjects and additional authors. Test the
following conditions:

OPACXSLTDetailsDisplay is blank and OPACPopupAuthorsSearch is off:

 - Clicking an additional author or subject link should directly trigger
   the correct search without an JavaScript error.

OPACXSLTDetailsDisplay is blank and OPACPopupAuthorsSearch is on:

 - Clicking an additional author or subject link should trigger a modal
   window where you can select options for searching. Checking boxes and
   submitting the form should perform the correct search.

OPACXSLTDetailsDisplay is set to "default" and OPACPopupAuthorsSearch is
on OR off:

 - Clicking an additional author or subject link should directly trigger
   the correct search without an JavaScript error.
   (OPACPopupAuthorsSearch only works with the non-XSL view).

For extra credit: Test with JavaScript disabled and confirm that link
still work correctly.

Signed-off-by: Frederic Demians <f.demians@tamil.fr>

I confirm the bug and the solution. Before various JS errors: "ReferenceError:
e is not defined" and "ReferenceError: showSubjects is not defined". After:
Comforting silence, and functionnality.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Works as described.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>