Bug 13375: Syspref search highlight styling broken in Chrome
This patch encloses the the syspref description/values mix in a <div>.
It doesn't introduce any regression on other browsers (like Firefox) and
has the side effect to make Chrome render the highlighted words on
syspref searches well.

This alternate patch uses a <div> instead of a <span> because it will
sometimes contain block-level elements. Follow the same plan:

To test:
- Search for 'facet' on the sysprefs
=> FAIL: verify it has styling problems in Chrome
- Apply the patch, reload
=> SUCCESS: styling is correct

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>