Bug 13449 - Syndetics TOC will not display in some cases
For reasons unknown, the table of contents data from Syndetics will not
display if there is no p element, even though within that if for the p
element there's a second one for the p element! This is neither good nor
necessary.

Test Plan:
1) Enable Syndetics
2) Find a record that should have a Syndetics TOC but doesn't
3) Apply this patch
4) Reload the page
5) Note the Syndetics TOC data now displays

Signed-off-by: Julius Fleschner <julius.fleschner@briarcliff.edu>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
I couldn't test but from reading the code this should not
have any negative consequences.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>